+++ Engine/Source/Runtime/Platform/OpenGL/OpenGLIndexBuffer.cpp
@@ -6,7 +6,8 @@
 namespace Hina
 {
 
-OpenGLIndexBuffer::OpenGLIndexBuffer(const uint32_t size, const Index *indices) : m_count(size / sizeof(Index)){
+OpenGLIndexBuffer::OpenGLIndexBuffer(const size_t size, const Index *indices)
+	: m_count(static_cast<uint32_t>(size / sizeof(Index))){
 	glGenBuffers(1, &m_renderID);
 	glBindBuffer(GL_ELEMENT_ARRAY_BUFFER, m_renderID);
 	glBufferData(GL_ELEMENT_ARRAY_BUFFER, size, indices, GL_STATIC_DRAW);
