+++ Engine/Source/Runtime/Platform/OpenGL/OpenGLVertexBuffer.cpp
@@ -8,7 +8,7 @@
 namespace Hina
 {
 
-OpenGLVertexBuffer::OpenGLVertexBuffer(const uint32_t size, const float *vertices) {
+OpenGLVertexBuffer::OpenGLVertexBuffer(const size_t size, const float *vertices) {
 	glGenBuffers(1, &m_renderID);
 	glBindBuffer(GL_ARRAY_BUFFER, m_renderID);
 	glBufferData(GL_ARRAY_BUFFER, size, vertices, GL_STATIC_DRAW);
